test-remote: Increase target timeout, move alerts-specific function into alerts test
